refactor code, init completeupgradewindow

+++ Assets/CodeBase/Data/EmployeeData.cs
@@ -1,5 +1,4 @@
 using System;
-using CodeBase.Gameplay.TableSystem;
 
 namespace CodeBase.Data
 {
@@ -10,8 +9,8 @@ namespace CodeBase.Data
         public string Id;
         public string Name;
         public int QualificationType;
-        public int Salary;
-        public int Profit;
+        public float Salary;
+        public float Profit;
         public bool IsWorking;
         public string TableId;
         public bool IsUpgrading;
